auth and adds working and uploaded

+++ app/Http/Controllers/User/PostsController.php
@@ -42,15 +42,35 @@ class PostsController extends Controller
             'title' => 'required|min:10|max:55',
             'body' => 'required|min:10|max:200',
             'category' => 'required',
-            'price' => 'required|numeric'
+            'price' => 'required|numeric',
+            'image_1' => 'image|max:1999|required'
         ]);
 
+        //Fandle file upload
+        if($request->hasFile('image_1')){
+            // Get filename with the extension
+            $filenameWithExt = $request->file('image_1')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just extension
+            $extension = $request->file('image_1')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            $request->image_1->move(public_path('user_images'), $fileNameToStore);
+        } else {
+            // Get just extension
+            $fileNameToStore = 'noimage.jpg';
+        }
+
+        // Create Adds
         $post = new Adds;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->category = $request->input('category');
         $post->price = $request->input('price');
         $post->user_id = auth()->user()->id;
+        $post->image_1 = $fileNameToStore;
         $post->save();
 
         return redirect('/user')->with('success', 'Skelbimas sukurtas...');
@@ -78,6 +98,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Adds::find($id);
+
+        //check for corect user , except Admin
+        if(auth()->user()->id !== $post->user_id && auth()->user()->type !== 'admin'){
+            return redirect('/skelbimai')->with('error', 'Neturi teisių...');
+        }
+
         return view('user.posts.edit')->with('post', $post);
     }
 
@@ -98,6 +124,11 @@ class PostsController extends Controller
         ]);
 
         $post = Adds::find($id);
+
+        if(auth()->user()->id !== $post->user_id && auth()->user()->type !== 'admin'){
+            return redirect('/skelbimai')->with('error', 'Neturi teisių...');
+        }
+
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->category = $request->input('category');
@@ -116,6 +147,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Adds::find($id);
+
+        //check for corect user , except Admin
+        if(auth()->user()->id !==$post->user_id && auth()->user()->type !== 'admin'){
+            return redirect('/skelbimai')->with('error', 'Neturi teisių...');
+        }
+
         $post->delete();
         return redirect('/user')->with('success', 'Skelbimas ištrintas...');
     }
